Separate Pages Into Components Fixes #31
